terraform_aws_lambda/requirements/dev.txt
-r ./base.txt

coverage
black[d]
boto3
moto
pytest
python-dotenv
